Write BioCro met files

Converts a met CF file to a model specific met file. The input files are called <in.path>/<in.prefix>.YYYY.cf

Usage

met2model.BIOCRO(
  in.path,
  in.prefix,
  outfolder,
  overwrite = FALSE,
  lat,
  lon,
  start_date,
  end_date,
  ...
)

Arguments

in.path

path on disk where CF file lives

in.prefix

prefix for each file

outfolder

location where model specific output is written

overwrite

logical: Write over any existing file of the same name? If FALSE, leaves the existing file untouched and skips to the next year.

lat, lon

Site latitude and longitude

start_date, end_date

Date range to convert. Each year will be written to a separate file

...

other arguments passed from PEcAn, currently ignored

Value

a dataframe of information about the written file

Author

Rob Kooper, David LeBauer