Run BioCro at a point

Run BioCro at a point

run.biocro(lat, lon, metpath, soil.nc = NULL, config = config,
  coppice.interval = 1)

Arguments

lat
latitude in decimal degrees

lon
longitude in decimal degrees

metpath
full path and name prefix of a csv file with hourly data in BioCro format, e.g. `/dir/met` if the files to be used are `/dir/met.2004.csv` and `dir/met.2005.csv'

soil.nc
full path and name of a netCDF file with soil data

config
full path and name of a config.xml file containing parameter values and configuration information for BioCro

coppice.interval
numeric, number of years between cuttings for coppice plant or perennial grass. Only used with BioCro 0.9; ignored when using later versions.

Value

output from one of the BioCro::*.Gro functions (determined by config$genus), as data.table object